Fixed: Template compiler script stripping characters
Changed the script to strip the last 5 characters from filenames (always '.html') to form the name of the template. Previously, it was removing the characters '.', 'h', 't', 'm', and 'l' from the end of filenames, which caused a problem with the name of my 'login-form.html' being passed to the template compiler as 'login-for' instead of 'login-form'.