Fixed: dashboard.css styles being overridden
Previously, JSPM was adding the bootstrap css link to the end of <head>, meaning that dashboard.css styles were being overridden by bootstrap.css. By changing bootstrap.css to the unminified version, the tag satisfies the dependency specified in bootstrap.js, so JSPM does not insert the css a second time (at the end of <head>).