WIP: ProjectDetailView
Increased view modularity: ProjectDetailView is nested inside ProjectLayoutView. It is responsible for showing the initial project view page, and making the request for WorkflowRuns. (The Workflows and Resources come from the subcollections on the initial /project/:uuid response.)

Also setting up infrastructure for Workflow, Resource, and WorkflowRun detail views which will be handled by ProjectLayoutView.